Wait for the node to start before scheduling  & Move jar to parent pom & optimize writer
