+++ seatunnel-engine/seatunnel-engine-storage/imap-storage-plugins/imap-storage-file/pom.xml
@@ -30,13 +30,6 @@
     <artifactId>imap-storage-file</artifactId>
     <name>SeaTunnel : Engine : Storage : IMap Storage Plugins : File</name>
 
-    <properties>
-        <hadoop-aliyun.version>3.0.0</hadoop-aliyun.version>
-        <hadoop-aws.version>3.1.4</hadoop-aws.version>
-        <aws.java.sdk.version>1.11.271</aws.java.sdk.version>
-        <netty-buffer.version>4.1.89.Final</netty-buffer.version>
-    </properties>
-
     <dependencies>
         <dependency>
             <groupId>org.apache.seatunnel</groupId>
@@ -74,34 +67,21 @@
         <dependency>
             <groupId>org.apache.hadoop</groupId>
             <artifactId>hadoop-aliyun</artifactId>
-            <version>${hadoop-aliyun.version}</version>
-            <scope>provided</scope>
         </dependency>
 
         <dependency>
-            <groupId>com.aliyun.oss</groupId>
-            <artifactId>aliyun-sdk-oss</artifactId>
-            <version>2.8.3</version>
-            <scope>provided</scope>
+            <groupId>net.minidev</groupId>
+            <artifactId>json-smart</artifactId>
         </dependency>
 
         <dependency>
             <groupId>org.apache.hadoop</groupId>
             <artifactId>hadoop-aws</artifactId>
-            <version>${hadoop-aws.version}</version>
-            <scope>provided</scope>
-        </dependency>
-        <dependency>
-            <groupId>com.amazonaws</groupId>
-            <artifactId>aws-java-sdk-bundle</artifactId>
-            <version>${aws.java.sdk.version}</version>
-            <scope>provided</scope>
         </dependency>
 
         <dependency>
             <groupId>io.netty</groupId>
             <artifactId>netty-buffer</artifactId>
-            <version>${netty-buffer.version}</version>
             <scope>provided</scope>
         </dependency>
