+++ tools/dependencies/known-dependencies.txt
@@ -22,8 +22,8 @@ protostuff-collectionschema-1.8.0.jar
 protostuff-core-1.8.0.jar
 protostuff-runtime-1.8.0.jar
 scala-library-2.11.12.jar
-seatunnel-jackson-2.3.1-SNAPSHOT-optional.jar
-seatunnel-guava-2.3.1-SNAPSHOT-optional.jar
+seatunnel-jackson-2.3.2-SNAPSHOT-optional.jar
+seatunnel-guava-2.3.2-SNAPSHOT-optional.jar
 slf4j-api-1.7.25.jar
 jsqlparser-4.5.jar
 animal-sniffer-annotations-1.17.jar
@@ -35,3 +35,6 @@ jsr305-1.3.9.jar
 jsr305-3.0.0.jar
 jsr305-3.0.2.jar
 listenablefuture-9999.0-empty-to-avoid-conflict-with-guava.jar
+accessors-smart-2.4.7.jar
+asm-9.1.jar
+json-smart-2.4.7.jar
